requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
streamlit>=1.10.0  # For dashboard
yfinance>=0.2.0    # For data collection 